requirements.txt
bump2version==1.0.1
discord.py==2.5.0
emoji==1.5.0
Pillow==9.5.0
pilmoji==2.0.1
requests==2.32.3
